logical assingment
